requirements.txt
# Core library for making HTTP requests
requests==2.31.0
# OpenAI Python client for API interaction
openai==0.28.1
# Library for creating PDF documents
reportlab==4.0.4
# Used to load environment variables from a .env file for local development
python-dotenv==1.0.0
